user profile Mordified

+++ app/Http/Controllers/UserProfileController.php
@@ -15,34 +15,65 @@ class UserProfileController extends Controller
 {
     public function index() {
     	$user = Auth::user(); //this is you active user logged in
-        // dd($user); use this to break and check your code
-        return response()->json($user);
+        return response()->json($user, 200);
     }
     public function all() {
+        $admins = [];
+        $residents = [];
+        $gatemans = [];
+
         $users = User::all();
-        return response()->json($users);
+        foreach ($users as $user) {
+            if($user->role == 0) {
+                array_push($admins, $user);
+            }else if($user->role == 1){
+                array_push($residents, $user);
+            }else if($user->role == 2) {
+                array_push($gatemans, $user);
+            }
+        }
+        $res['status']    = true;
+        $res['admins']    = $admins;
+        $res['residents'] = $residents;
+        $res['gatemans']  = $gatemans;
+        return response()->json($res, 200);
+    }
+
+    public function showOneAdmin($id) {
+        $user = User::find($id);
+        if($user->role == 0) {
+            $res['status'] = true;
+            $res['message'] = 'Admin found';
+            $res['admin'] = $user;
+            return response()->json($res, 200);
+        }else {
+            $res['status'] = false;
+            $res['message'] = 'Admin not found';
+            return response()->json($res, 404);
+        }
     }
 
     public function show($id) {
         $user = User::find($id);
-        return response()->json($user);
+        if($user->role == 1 || $user->role == 2) {
+            $res['status'] = true;
+            $res['message'] = 'User found';
+            $res['user'] = $user;
+            return response()->json($res, 200);
+        }else {
+            $res['status'] = false;
+            $res['message'] = 'User not found';
+            return response()->json($res, 404);
+        }
     }
 
-    public function role($role) {
-        $users = DB::table('users')->select('*')->where('user.role', parseInt($role))->get();
+
+    public function role($role_id) {
+        $users = User::where('role', $role_id)->get();
         return response()->json($users);
 
     }
 
-    public function password(Request $request) {
-        $user = Auth::user();
-      
-        $user->password = Hash::make($request->input('password'));
-        $user->save();
-        return response()->json($user);
-
-    }
-  
     public function update(Request $request) { 
         $user = Auth::user();
         $this->validate($request, [
@@ -82,11 +113,45 @@ class UserProfileController extends Controller
         }
    }
 
+    public function password(Request $request) {
+        $user = Auth::user();
+
+        $this->validate($request, [
+            'old_password'=> 'required|string',
+            'password' => 'required|min:8|different:old_password|confirmed'
+        ]);
+         //start temporay transaction
+        DB::beginTransaction();
+        try{
+                $user->password = Hash::make($request->input('password'));
+                $user->save();
+
+                 //if operation was successful save commit save to database
+                DB::commit();
+                $res['status'] = true;
+                $res['message'] = 'Password Changed Successfully!';
+                return response()->json($res, 200);
+            }catch(\Exception $e) {
+
+                //rollback what is saved
+                DB::rollBack();
+                $res['status'] = false;
+                $res['message'] = 'Password Update unsuccessful: An error occured, please try again!';
+                return response()->json($res, 501);
+        }
+
+
+    }
+
     public function destroy() {
         $user = Auth::user();
-        $user->delete();
-
-        $reponse = array('response' => 'Item deleted', 'success' => true);
-        return $response;
+        if($user) {
+            $user->delete();
+            $res['message'] = 'User deleted successfully';
+            return response()->json($res, 200);
+        }else {
+            $res['message'] = 'User unsuccessfully, user not found or an error occured, please try again!';
+            return response()->json($res, 501);
+        }
     }
 }
